# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KuruOps API
  version: "0.1.0"
  description: >
    REST API for KuruOps, a single-tenant SOC alert/incident-management
    platform (Go backend, PostgreSQL, React frontend). This document covers
    every route mounted by internal/httpserver/router.go: the unauthenticated
    /auth/** login flows, /healthz and /metrics, and the JWT-protected
    /api/v1/** surface (alerts, incidents, playbooks, tags, uploads,
    dashboard, live event stream, and admin-only /settings/** configuration).
    See the repository root README for product context, deployment, and the
    default-admin-first-login flow; see TROUBLESHOOTING.md for operational
    issues. This spec favors completeness of shape (paths, params, request/
    response bodies, status codes) over exhaustive field-level validation
    constraints.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/incidents:
    get:
      tags: [Incidents]
      summary: List incidents
      description: Requires the "incidents" resourceAccess capability.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: severity
          in: query
          schema: { $ref: '#/components/schemas/Severity' }
        - name: priority
          in: query
          schema: { $ref: '#/components/schemas/IncidentPriority' }
        - name: phase
          in: query
          schema: { $ref: '#/components/schemas/IncidentPhase' }
        - name: tag
          in: query
          schema: { type: string }
        - name: q
          in: query
          description: Full-text search against title and description.
          schema: { type: string }
        - name: since
          in: query
          description: RFC3339 lower bound on openedAt.
          schema: { type: string, format: date-time }
        - name: commanderId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Matching incidents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Incident' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Incidents]
      summary: Create an incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                severity: { $ref: '#/components/schemas/Severity' }
                priority: { $ref: '#/components/schemas/IncidentPriority' }
                assigneeIds:
                  type: array
                  items: { type: string, format: uuid }
                tags:
                  type: array
                  items: { type: string }
              required: [title, severity, priority]
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v1/incidents/{id}:
    get:
      tags: [Incidents]
      summary: Get an incident
      description: Includes roles (unlike the list endpoint).
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        "200":
          description: The incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalError'
...
  /api/v1/settings/incident-sla:
    get:
      tags: [Settings - Incident SLA]
      summary: List incident SLA policies
      responses:
        "200":
          description: SLA policies
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IncidentSLAPolicy' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Settings - Incident SLA]
      summary: Create or update the SLA policy for a (severity, priority) pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                severity: { $ref: '#/components/schemas/Severity' }
                priority: { $ref: '#/components/schemas/IncidentPriority' }
                dueWithinMinutes: { type: integer }
              required: [severity, priority, dueWithinMinutes]
      responses:
        "200":
          description: Policy saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentSLAPolicy' }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v1/settings/incident-sla/{id}:
    delete:
      tags: [Settings - Incident SLA]
      summary: Delete an SLA policy
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalError'

  # ------------------------------------------- Settings: Escalation Policies
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Required on every /api/v1/** route. /auth/** is intentionally
        unauthenticated (that's how a session is obtained in the first
        place), and /healthz and /metrics are unauthenticated for
        infrastructure health checks and scraping. A caller whose
        mustChangePassword flag is still set can reach nothing under
        /api/v1/** except POST /api/v1/account/change-password (see
        RequirePasswordChanged), receiving 403 on everything else until the
        password is changed.
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size, 1-200. Invalid or out-of-range values fall back to the server default (50).
      schema: { type: integer, minimum: 1, maximum: 200 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0 }
...
    IncidentId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Malformed request body, path, or query parameter
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Unauthorized:
      description: Missing, malformed, or expired bearer token
      content:
        text/plain:
          schema:
            type: string
            description: >-
              Plain-text error message, issued by the auth middleware via
              http.Error -- not the handler-level JSON error shape used
              elsewhere in this API.
    Forbidden:
      description: Authenticated, but the caller's role/resourceAccess doesn't permit this action (or the account still must change its password)
      content:
        text/plain:
          schema:
            type: string
            description: >-
              Plain-text error message, issued by the auth middleware via
              http.Error -- not the handler-level JSON error shape used
              elsewhere in this API.
    NotFound:
      description: No resource with that id (or not visible under the caller's tag scope)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: Standard handler-level error shape, written by handlers.writeError.
      properties:
        error:
          type: string
      required: [error]
...
    Severity:
      type: string
      enum: [critical, high, medium, low, informational]
...
    IncidentPhase:
      type: string
      enum: [new, detection_analysis, containment, eradication, recovery, post_incident]
    IncidentPriority:
      type: string
      enum: [p1, p2, p3, p4]
    IncidentRole:
      type: string
      enum: [commander, technical_lead, incident_handler, communications_lead, privacy_officer]
...
    UserSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
...
    Incident:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        severity: { $ref: '#/components/schemas/Severity' }
        priority: { $ref: '#/components/schemas/IncidentPriority' }
        phase: { $ref: '#/components/schemas/IncidentPhase' }
        assignees:
          type: array
          items: { $ref: '#/components/schemas/UserSummary' }
        roles:
          type: array
          description: Only populated on GET of a single incident, not on list.
          items: { $ref: '#/components/schemas/IncidentRoleAssignment' }
        tags:
          type: array
          items: { type: string }
        slaDueAt: { type: string, format: date-time, nullable: true }
        slaBreached: { type: boolean }
        openedAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        latestAnalysis:
          type: string
          nullable: true
          description: Most recent completed "Analyze with AI" result text -- only set once latestAnalysisStatus is "completed".
        latestAnalysisStatus:
          type: string
          nullable: true
          enum: [running, paused, completed, failed]
          description: Null if no analysis has ever been requested for this incident.
        latestAnalysisError:
          type: string
          nullable: true
          description: Set only when latestAnalysisStatus is "failed".
      required: [id, tenantId, title, description, severity, priority, phase, assignees, roles, tags, slaBreached, openedAt, createdAt, updatedAt]
    IncidentRoleAssignment:
      type: object
      properties:
        role: { $ref: '#/components/schemas/IncidentRole' }
        user: { $ref: '#/components/schemas/UserSummary' }
...
    IncidentSLAPolicy:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        severity: { $ref: '#/components/schemas/Severity' }
        priority: { $ref: '#/components/schemas/IncidentPriority' }
        dueWithinMinutes: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...